feat: finalize.callback called with result
As `finalize` is called with `triggerHandler` we can read back the result from the handlers and forward to the `finalize` callback. This means we can now do this:

```javascript
var Widget = require("mu-jquery-widget/widget");
var W = Widget.extend({
  "on/finalize": function ($event, finalized) {
    finalized(function(result) {
      console.log(result === "result" // true
    });
    return "result";
  }
 });

var w = new W($("<div>", "ns");
w.remove();
```

BREAKING CHANGE: The removal of event handlers in `finalize` is removed in this commit as it's just as easy to do `w.off()` on your own when needed.